+++ CHANGELOG.md
@@ -15,3 +15,7 @@
 
 ## v1.0.1
 1. Improved the Ruby Gem Specification to depend on [unidom-common](https://github.com/topbitdu/unidom-common) v1.0
+
+## v1.1
+1. As Source Standard concern
+2. As Target Standard concern